Update Test-Simple to the latest dev release
This updates Test-Simple to the latest dev release. This dev release is
based on Test2 now. This is part of my grant work.

The idea where is to get this into blead for testing and smoking
purposes. Putting it here insures more eyes will see it.

Test-Simple now depends on Test, so that has been included.

I also needed to tweak ext/XS-APItest/t/svpeek.t to adjust for how $!
and $@ are handled now by the testing tools.