XXX Tests Use different algorithm in mem_collxfrm() to handle embedded NULs
Perl uses strxfrm() to handle collation.  This C library function
expects a NUL-terminated input string.  But Perl accepts interior NUL
charaters, so something has to happen.

Until this commit, what happened was that each NUL-terminated
sub-segment would be individually passed to strxfrm(), with the results
concatenated together to form the transformation of the whole string
with NULs ignored.  But this isn't guaranteed to give good results, as
strxfrm() is highly context sensitive, and needs the whole string, not
segments, to work properly.  The way strxfrm() works, more or less,  is
that it returns a string consisting of the primary weights, in order,
of the characters of the input, concatenated with the secondary weights,
and so on.  Giving strxfrm() only substrings defeats this.

Another possibility would be to just remove the NULs before transforming
the string.  The problem with this method is that it screws up the
context.  In some locales, two adjacent characters can behave
differently than if they were separated.

What this commit does is to change to replace each NUL with a \001.
\001 is almost certainly going to behave like we expect a NUL would if
it were legal.  Just about every locale treats low code points as
controls, to be ignored in at least primary weighting.

And this method gives the expected sort order.  This is because perl
uses the original strings as a tie breaker.  So, given two strings, one
that originally had \001, and one that differed only in that it had \000
instead, they both will get the same transformation, so will sort equal
there, but the tie breaker will cause the one with NULs to sort first.

As stated in the comments, we could go through the first 256 code points
to determine the lowest collating one, instead of assuming it is \001.
But this is a lot of work (UTF-8ness must be considered) and it will be
extremely rare that the answer isn't going to be \001.