locale.c: XXX not so aggressive guess incre
On platforms where  strxfrm() is not well-behaved and it fails because
it needs a larger buffer, prior to this commit, the size was doubled
before trying again.  This could require a lot of memory on large
inputs.  This commit changes it so it is not so aggressive.  I think the
size prediction is better due to a recent commit, and there isn't much
of a downside in not gobbling up memory so fast (although the excess is
soon freed).